--- .idea/runConfigurations/Python_tests_for_test_integration_IntegrationTestPlaying_test_walk_and_kick_right.xml
@@ -1,17 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Python tests for test_integration.IntegrationTestPlaying.test_walk_and_kick_right" type="tests" factoryName="Autodetect" nameIsGenerated="true">
-    <module name="soccerbot" />
-    <option name="INTERPRETER_OPTIONS" value="" />
-    <option name="PARENT_ENVS" value="true" />
-    <option name="SDK_HOME" value="" />
-    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
-    <option name="IS_MODULE_SDK" value="true" />
-    <option name="ADD_CONTENT_ROOTS" value="true" />
-    <option name="ADD_SOURCE_ROOTS" value="true" />
-    <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py" />
-    <option name="_new_additionalArguments" value="&quot;-s&quot;" />
-    <option name="_new_target" value="&quot;test_integration.IntegrationTestPlaying.test_walk_and_kick_right&quot;" />
-    <option name="_new_targetType" value="&quot;PYTHON&quot;" />
-    <method v="2" />
-  </configuration>
-</component>
